+++ R/helper_elimination_game.R
@@ -14,81 +14,33 @@ new_min_sp_triang <- function(x, nlvls) {
   structure(list(x = x, nlvls =  nlvls[dimnames(x)[[1]]]), class = c("min_sp_triang", "list"))
 }
 
-new_minimal_triang <- function(x) {
-    structure(list(x = x), class = c("minimal_triang", "list"))
-}
-
-# new_sparse_triang can be found in tmp_helper_sparse_triang.
-
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#          HELPERS TO FIND NEXT ELIMINATION NODE 
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-# new_node_to_eliminate_sparse <- function(spt, x) {
-#   # x: The submatrix during elimination
-#   nnzc <- .map_dbl(1:ncol(x), function(k) {
-
-#     new_node <- k
-#     nei_idx <- unname(which(x[, new_node] == 1L))
-#     nei_x   <- x[nei_idx, nei_idx]
-#     length(nei_idx)
-
-#     # which messages and which non_allocated can be put into (new_node, nei_idx)
-#     new_prime_chr <- colnames(x)[c(new_node, nei_idx)]
-
-#     na_cpts_idx <- .map_lgl(spt$tmp_potentials$non_allocated_cpts, function(cpt) {
-#       all(names(cpt) %in% new_prime_chr)
-#     })
-
-#     msg_cpts_idx <- .map_lgl(spt$tmp_potentials$flawed_root_msg, function(cpt) {
-#       all(names(cpt) %in% new_prime_chr)
-#     })
-
-#     na_cpts  <- spt$tmp_potentials$non_allocated_cpts[na_cpts_idx]
-#     msg_cpts <- spt$tmp_potentials$flawed_root_msg[msg_cpts_idx]
-#     all_cpts <- c(na_cpts, msg_cpts)
-
-#     1)
-#     sparsity_ <- function(x) 1 - ncol(x) / prod(.map_int(sparta::dim_names(x), length))
-#     2) Mark msg's
-#     2.5) Gang "samme" potentialer sammen
-#     3) Log final clique pot size vs. estimated
-#     4) Use the updated JT in jt().
-#     5) Findings / evidence
-
-#     statespace_vars <- spt$dns[unique(unlist(lapply(all_cpts, names)))]
-#     statespace <- prod(.map_int(statespace_vars, length))
-#     sparsity_  <- prod(.map_dbl(all_cpts, function(x) {
-#       sparsity_(x)
-#     }))
-
-#     # Estimate of number of non-zero elements in the product
-#     sparsity_ * statespace * length(statespace_vars)
-
-#     # Sanity check
-#     # Reduce(sparta::mult, all_cpts)
-#   })
-
-#   new_node     <- which.min(nnzc)
-#   nei_idx      <- unname(which(x[, new_node] == 1L))
-#   x_nei        <- x[nei_idx, nei_idx]
-#   nei_complete <- sum(x_nei) == length(nei_idx) * (length(nei_idx) - 1)
-
-#   # INSPECTION
-#   # ----------
-#   # browser()
-#   # igraph::graph_from_adjacency_matrix(x, "undirected") %>% plot(vertex.size = .1, vertex.label = NA)
-#   # colnames(x)[new_node]
-#   # ----------
-  
-#   return(list(x_nei = x_nei, nei_idx = nei_idx, nei_complete = nei_complete, a = new_node))
+# new_evidence_triang <- function(x, evidence_nodes) {
+#   # Note: Evidence nodes must be ordered according to their priority
+#   bools <- vector("logical", length = ncol(x))
+#   if (neq_empt_chr(evidence_nodes)) {
+#     # Test if evidence_are valid
+#     stopifnot(all(evidence_nodes %in% colnames(x)))
+#     bools[match(evidence_nodes, colnames(x))] <- TRUE
+#   }
+#   structure(list(x = x, x_nei_lookup = x, evidence_nodes = bools), class = c("evidence_triang", "list"))
 # }
+
+new_evidence_triang <- function(x, evidence_nodes) {
+  # Note: Evidence nodes must be ordered according to their priority
+  bools <- vector("logical", length = ncol(x))
+  if (neq_empt_chr(evidence_nodes)) {
+    # Test if evidence_are valid
+    stopifnot(all(evidence_nodes %in% colnames(x)))
+    bools[match(evidence_nodes, colnames(x))] <- TRUE
+  }
+  structure(list(x = x, evidence_nodes = bools), class = c("evidence_triang", "list"))
+}
 
 
 new_node_to_eliminate_min_nei <- function(x) {
   # Minimum-neibor/minimum size elimination:
   # Minimizing the number of fillins by choosing vertices that
-  # hopefully simplicial
+  # are hopefully simplicial
   lookup_order <- as.integer(names(sort(structure(apply(x, 2L, sum), names = 1:ncol(x)))))
   new_node_idx <- lookup_order[1L]
   nei     <- x[, new_node_idx, drop = TRUE]
@@ -150,17 +102,87 @@ new_node_to_eliminate_min_sp <- function(x, nlvls) {
 }
 
 
+# new_node_to_eliminate_evidence <- function(x, x_nei_lookup, evidence_nodes) {
+
+#   n_new_fills <- .map_dbl(1:ncol(x), function(k) {
+#     nei_idx_k <- which(x[, k] == 1L)
+#     all_edges <- length(nei_idx_k) * (length(nei_idx_k) - 1L) / 2
+#     existing_edges <- sum(x[nei_idx_k, nei_idx_k]) / 2L
+#     all_edges - existing_edges
+#   })
+  
+#   F_ <- which(n_new_fills == min(n_new_fills))
+#   E  <- which(evidence_nodes == TRUE)
+#   H  <- setdiff(F_, E)
+
+#   new_node <- if (neq_empt_int(H)) {
+#     if (neq_empt_int(E)) {
+#       H_neis   <- lapply(E, function(e) which(H %in% which(x_nei_lookup[, e] == 1L)))
+#       # H_neis   <- lapply(E, function(e) which(F_ %in% which(x_nei_lookup[, e] == 1L)))
+#       H_neis   <- Filter(neq_empt_int, H_neis)
+#       new_idx <- if (neq_empt_lst(H_neis)) {
+#         H_neis[[1]][1]
+#       } else {
+#         1L
+#       }
+#       H[new_idx]
+#     } else { # E empty
+#       F_[1]
+#     }
+#   } else {   # H empty
+#     F_[1]
+#   }
+  
+#   nei_idx      <- which(x[, new_node] == 1L)
+#   x_nei        <- x[nei_idx, nei_idx, drop = FALSE]
+#   nei_complete <- sum(x_nei) == length(nei_idx) * (length(nei_idx) - 1)
+
+#   return(list(x_nei = x_nei, nei_idx = nei_idx, nei_complete = nei_complete, a = new_node))
+# }
+
+
+new_node_to_eliminate_evidence <- function(x, evidence_nodes) {
+
+  n_new_fills <- .map_dbl(1:ncol(x), function(k) {
+    nei_idx_k <- which(x[, k] == 1L)
+    all_edges <- length(nei_idx_k) * (length(nei_idx_k) - 1L) / 2
+    existing_edges <- sum(x[nei_idx_k, nei_idx_k]) / 2L
+    all_edges - existing_edges
+  })
+  
+  F_ <- which(n_new_fills == min(n_new_fills))
+  E  <- which(evidence_nodes == TRUE)
+  H  <- setdiff(F_, E)
+  W  <- intersect(F_, E)
+  
+  new_node <- if (neq_empt_int(E)) {
+    if (neq_empt_int(H)) {
+      neis <- lapply(E, function(e) which(H %in% which(x[, e] == 1L)))
+      neis <- Filter(neq_empt_int, neis)
+      if (neq_empt_lst(neis)) H[neis[[1]][1]] else F_[1]      
+    } else {
+      neis <- lapply(E, function(e) which(W %in% which(x[, e] == 1L)))
+      neis <- Filter(neq_empt_int, neis)
+      if (neq_empt_lst(neis)) W[neis[[length(neis)]][1]] else F_[1]      
+    }
+
+  } else {
+    F_[1]
+  }
+  
+  nei_idx      <- which(x[, new_node] == 1L)
+  x_nei        <- x[nei_idx, nei_idx, drop = FALSE]
+  nei_complete <- sum(x_nei) == length(nei_idx) * (length(nei_idx) - 1)
+
+  return(list(x_nei = x_nei, nei_idx = nei_idx, nei_complete = nei_complete, a = new_node))
+}
+
+
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #               ELIMINATION GAME ENGINE
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 elim_game <- function(obj) {
   # x: adjacency matrix
-
-  # x <- if (inherits(obj, "sparse_triang")) {
-  #   obj$flawed_root_graph
-  # } else {
-  #   obj$x
-  # }
 
   x <- obj$x
 
@@ -170,6 +192,7 @@ elim_game <- function(obj) {
   # triangulation set
   fill_edges <- list()
 
+  # elimination ordering
   alpha      <- vector("integer", length = ncol(x))
   alpha_iter <- 1L
   
@@ -183,25 +206,23 @@ elim_game <- function(obj) {
       new_node_to_eliminate_min_fill(x)
     } else if (inherits(obj, "min_sp_triang")) {
       new_node_to_eliminate_min_sp(x, obj$nlvls)
-    } else if (inherits(obj, "minimal_triang")) {
-      new_node_to_eliminate_min_fill(x)
+    } else if (inherits(obj, "evidence_triang")) {
+      out <- new_node_to_eliminate_evidence(x, obj$evidence_nodes)
+      # Eliminate corresponing idx in evidence nodes and update neighborhood of evidence nodes
+      obj$evidence_nodes <- obj$evidence_nodes[-out$a]
+      # obj$x_nei_lookup   <- obj$x_nei_lookup[-out$a, -out$a, drop = FALSE]
+      out
     }
-
-    # else if (inherits(obj, "sparse_triang")) {
-    #   new_node_to_eliminate_sparse(spt, x)
-    # }
     
     xa        <- X$a
     nc        <- X$nei_complete
     
     if (!nc) { # append new fill_edges
-
       x_nei   <- X$x_nei
       nei_idx <- X$nei_idx
       nn      <- ncol(x_nei)
 
       for (k in 1:(nn-1)) {
-
         x_nei_k <- x_nei[, k, drop = TRUE]
         non_adj <- nei_idx[which(x_nei_k[(k+1):nn] == 0L) + k]
         fills   <- lapply(non_adj, function(a) c(nei_idx[k], a))
@@ -226,8 +247,6 @@ elim_game <- function(obj) {
     
     x                 <- x[-xa, -xa, drop = FALSE]
     x_orig_col_idx    <- x_orig_col_idx[-xa]
-
-
   }
 
    # should only ontain 1 value now